+++ springwolf-add-ons/springwolf-common-model-converters/src/main/java/io/github/stavshamir/springwolf/addons/common_model_converters/configuration/CommonModelConvertersAutoConfiguration.java
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
-package io.github.stavshamir.springwolf.common_converters.configuration;
+package io.github.stavshamir.springwolf.addons.common_model_converters.configuration;
 
-import io.github.stavshamir.springwolf.common_converters.converters.monetaryamount.MonetaryAmountConverter;
+import io.github.stavshamir.springwolf.addons.common_model_converters.converters.monetaryamount.MonetaryAmountConverter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
