+++ springwolf-add-ons/springwolf-common-model-converters/src/test/java/io/github/stavshamir/springwolf/addons/common_model_converters/converters/MonetaryAmountConverterTest.java
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
-package io.github.stavshamir.springwolf.common_converters.converters;
+package io.github.stavshamir.springwolf.addons.common_model_converters.converters;
 
-import io.github.stavshamir.springwolf.common_converters.converters.monetaryamount.MonetaryAmountConverter;
+import io.github.stavshamir.springwolf.addons.common_model_converters.converters.monetaryamount.MonetaryAmountConverter;
 import io.swagger.v3.core.converter.AnnotatedType;
 import io.swagger.v3.core.converter.ModelConverters;
 import io.swagger.v3.oas.models.media.Schema;
